Isolate stage/job/sql buffer purge
We observed asynchronous behaviour of event method calling. This means
it is not safe enough to assume that a job end event arrives strictly
after a stage end. This represents a problem because buffers could be
left with orphan stage entries if only purged on end job arrival.

In this change we made them independent. The only assumption is that:
- job start arrives before job end
- sql start arrives before sql end